small refactorings to fix CLI import errors for replicator
